+++ NEWS.md
@@ -1,5 +1,6 @@
 ### Unreleased
 
+* Added support for setting custom properties on the project (#2903)
 * Removed Space and Ctrl+Space shortcuts from Layers view to avoid conflict with panning (#3672)
 * Scripting: Added API for editing tile layers using terrain sets (with a-morphous, #3758)
 * Fixed object preview position with parallax factor on group layer (#3669)